Added comments

+++ Core/Inc/bmp280_read.h
@@ -19,7 +19,7 @@ void delay_ms(uint32_t period_ms);
 int8_t i2c_reg_write(uint8_t i2c_addr, uint8_t reg_addr, uint8_t *reg_data, uint16_t length);
 int8_t i2c_reg_read(uint8_t i2c_addr, uint8_t reg_addr, uint8_t *reg_data, uint16_t length);
 int8_t BMP280_init(void);
-int BMP280_TempRead(double *temperature);
-int BMP280_PressRead(double *pressure);
+int8_t BMP280_TempRead(double *temperature);
+int8_t BMP280_PressRead(double *pressure);
 double getTemp(double temperature);
 #endif /* INC_BMP280_READ_H_ */
